requirements.txt
# Core LangChain
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
langchain-anthropic==0.2.0

# Vector Stores
chromadb==0.5.0
pinecone-client==5.0.0
pgvector==0.3.0

# Document Processing
pypdf==4.0.0
python-docx==1.0.1
tiktoken==0.7.0

# Utilities
python-dotenv==1.0.0
pydantic==2.0.0

# Development
jupyter==1.0.0
ipykernel==6.29.0

# Observability
langsmith==0.1.0

# API & Deployment
fastapi==0.115.0
uvicorn==0.30.0
streamlit==1.38.0

# Data processing (leverage your existing skills)
pandas==2.2.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.9
